block does something

+++ backend/src/filetObish/Comment.java
@@ -4,5 +4,5 @@ public abstract class Comment {
 	//there is content and id
 	public String content;
 	public long id;
-	public abstract void report(int id);
+	public abstract void report(long id);
 }
